Correcting bug when game had no estimated infos
When a game had no estimated duration and/or estimated winning team and
we are consulting a summoner's page that also have no 'data' information
about that game, it was considered as lost and during 0min.
This commit fixes the bug by computing the estimation and saving it in
the game table when the page tries to display a game that is 0min long.
The game still counts as 0min/loss for the actual page's statistics,
though.